batch_inference/requirements.txt
kafka_python==2.1.5
loguru==0.7.3
pymongo==4.12.0
vllm==0.8.4
